% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbons.R
\docType{data}
\name{gibbons}
\alias{gibbons}
\title{Gibbon movement locations from Khao Yai National Park}
\format{
An \code{sf} object containing 18,098 point locations.
}
\usage{
gibbons
}
\description{
Movement locations recorded at approximately 5-m intervals
throughout 2003 for a group of white-handed gibbons
(\emph{Hylobates lar}) in Khao Yai National Park, Thailand.
}
\details{
The dataset contains 18,098 spatial locations represented
as an \code{sf} point object. Coordinates are provided in UTM
zone 47N (EPSG:32647).

The dataset was originally used to examine short-term
core-area use in white-handed gibbons.
}
\examples{
data(gibbons)

grid_hr(
  gibbons,
  cell_area = 5000,
  fill_by = "radial",
  show_legend = FALSE,
  title = "Radial space-use intensity of a gibbon group
  (1 hex = 0.5 ha)"
)
}
\references{
Asensio, N., Brockelman, W. Y., Malaivijitnond, S. &
Reichard, U. H. (2014).
White-handed Gibbon (\emph{Hylobates lar}) Core Area Use Over
a Short-Time Scale. \emph{Biotropica}, 46, 461-469.
}
\keyword{datasets}
