% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider_monkeys.R
\docType{data}
\name{spider_monkeys}
\alias{spider_monkeys}
\title{Spider monkey subgroup locations}
\format{
An \code{sf} object containing spider monkey subgroup locations.
}
\source{
Asensio, N., Murillo-Chacon, E., Schaffner, C. M. & Aureli, F. (2017).
\doi{10.1111/btp.12441}
}
\usage{
spider_monkeys
}
\description{
Spatial locations of spider monkey subgroups recorded at approximately
30-minute intervals from 2005 to 2008 in a regenerating tropical forest.
Coordinates are provided in UTM zone 16N using the WGS84 datum
(EPSG:32616).
}
\details{
The dataset was used to investigate the effects of roads on spider
monkey home-range size and mobility in a heterogeneous regenerating
forest.
}
\examples{
data(spider_monkeys)

grid_hr(
  spider_monkeys,
  cell_area = 10000,
  fill_by = "radial",
  title = "Radial space-use intensity of a spider monkey group (1 ha hexagons)"
)
}
\references{
Asensio, N., Murillo-Chacon, E., Schaffner, C. M. & Aureli, F. (2017).
The effect of roads on spider monkeys' home range and mobility in a
heterogeneous regenerating forest. \emph{Biotropica}.
\doi{10.1111/btp.12441}
}
\keyword{datasets}
